+++ src/loader-components/_Badges.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import styles from './NpmBadge.module.scss';
+import styles from './Badges.module.scss';
 
 const cx = classNames.bind(styles);
 
@@ -10,6 +10,10 @@ const propTypes = {
    * component name.
    */
   name: PropTypes.string.isRequired,
+  /**
+   * Source code url
+   */
+  src: PropTypes.string,
   /**
    * url to link to
    */
@@ -20,8 +24,15 @@ const propTypes = {
   version: PropTypes.string.isRequired,
 };
 
-const NpmBadge = ({ name, url, version }) => (
-  <div className={cx('badge-container')}>
+const Badges = (props) => {
+  const {
+    src,
+    name,
+    url,
+    version,
+  } = props;
+
+  const npmBadge = (
     <a className={cx('badge')} href={url || `https://www.npmjs.org/package/${name}/v/${version}`}>
       <span className={cx('badge-name')}>
         {url ? 'package' : 'npm'}
@@ -30,9 +41,27 @@ const NpmBadge = ({ name, url, version }) => (
         {`v${version}`}
       </span>
     </a>
-  </div>
-);
+  );
 
-NpmBadge.propTypes = propTypes;
+  const srcCodeBadge = (src) ? (
+    <a className={cx('badge')} href={src}>
+      <span className={cx('badge-name')}>
+        github
+      </span>
+      <span className={cx('badge-version')}>
+        source
+      </span>
+    </a>
+  ) : undefined;
 
-export default NpmBadge;
+  return (
+    <div className={cx('badge-container')}>
+      { npmBadge }
+      { srcCodeBadge }
+    </div>
+  );
+};
+
+Badges.propTypes = propTypes;
+
+export default Badges;